Move enums and interfaces closer to logic owners
