+++ ulox/ulox.core/Package/Runtime/Engine/CallFrame.cs
@@ -1,5 +1,11 @@
 ﻿namespace ULox
 {
+    public enum NativeCallResult
+    {
+        SuccessfulExpression,
+        Failure
+    }
+    
     public struct CallFrame
     {
         public delegate NativeCallResult NativeCallDelegate(Vm vm);
