--- ulox/ulox.core/Package/Runtime/Enums/ClosureType.cs
@@ -1,8 +1,0 @@
-﻿namespace ULox
-{
-    public enum ClosureType : byte
-    {
-        Closure,
-        UpValueInfo,
-    }
-}
