--- ulox/ulox.core/Package/Runtime/Enums/FunctionType.cs
@@ -1,11 +1,0 @@
-﻿namespace ULox
-{
-    public enum FunctionType
-    {
-        Script,
-        Function,
-        Method,
-        Init,
-        TypeDeclare,
-    }
-}
