--- ulox/ulox.core/Package/Runtime/Enums/InterpreterResult.cs
@@ -1,8 +1,0 @@
-﻿namespace ULox
-{
-    public enum InterpreterResult
-    {
-        OK,
-        YIELD,
-    }
-}
