--- ulox/ulox.core/Package/Runtime/Enums/NativeCallResult.cs
@@ -1,8 +1,0 @@
-﻿namespace ULox
-{
-    public enum NativeCallResult
-    {
-        SuccessfulExpression,
-        Failure
-    }
-}
