--- ulox/ulox.core/Package/Runtime/Enums/NativeType.cs
@@ -1,8 +1,0 @@
-﻿namespace ULox
-{
-    public enum NativeType : byte
-    {
-        List,
-        Dynamic,
-    }
-}
