--- ulox/ulox.core/Package/Runtime/Enums/Precedence.cs
@@ -1,17 +1,0 @@
-﻿namespace ULox
-{
-    public enum Precedence
-    {
-        None,
-        Assignment,
-        Or,
-        And,
-        Equality,
-        Comparison,
-        Term,
-        Factor,
-        Unary,
-        Call,
-        Primary,
-    }
-}
