--- ulox/ulox.core/Package/Runtime/Enums/PushValueOpType.cs
@@ -1,10 +1,0 @@
-﻿namespace ULox
-{
-    public enum PushValueOpType : byte
-    {
-        Null,
-        Bool,
-        Byte,
-        Bytes,
-    }
-}
