--- ulox/ulox.core/Package/Runtime/Enums/TestOpType.cs
@@ -1,11 +1,0 @@
-﻿namespace ULox
-{
-    public enum TestOpType : byte
-    {
-        TestFixtureBodyInstruction,
-        TestCase,
-        CaseStart,
-        CaseEnd,
-        TestSetEnd,
-    }
-}
