--- ulox/ulox.core/Package/Runtime/Enums/ValidateOp.cs
@@ -1,8 +1,0 @@
-﻿namespace ULox
-{
-    public enum ValidateOp : byte
-    {
-        Meets,
-        Signs,
-    }
-}
