--- ulox/ulox.core/Package/Runtime/Enums/ValueType.cs
@@ -1,18 +1,0 @@
-﻿namespace ULox
-{
-    public enum ValueType : byte
-    {
-        Null,
-        Double,
-        Bool,
-        String,
-        Chunk,
-        NativeFunction,
-        Closure,
-        Upvalue,
-        UserType,
-        Instance,
-        BoundMethod,
-        Object,
-    }
-}
