--- ulox/ulox.core/Package/Runtime/Interfaces/ICompilette.cs
@@ -1,9 +1,0 @@
-﻿namespace ULox
-{
-    public interface ICompilette
-    {
-        TokenType MatchingToken { get; }
-
-        void Process(Compiler compiler);
-    }
-}
