--- ulox/ulox.core/Package/Runtime/Interfaces/IParseRule.cs
@@ -1,9 +1,0 @@
-﻿namespace ULox
-{
-    public interface IParseRule
-    {
-        Precedence Precedence { get; }
-        void Prefix(Compiler compiler, bool canAssign);
-        void Infix(Compiler compiler, bool canAssign);
-    }
-}
