--- ulox/ulox.core/Package/Runtime/Interfaces/IScannerTokenGenerator.cs
@@ -1,9 +1,0 @@
-﻿namespace ULox
-{
-    public interface IScannerTokenGenerator
-    {
-        bool DoesMatchChar(char ch);
-
-        void Consume(Scanner scanner);
-    }
-}
