--- ulox/ulox.core/Package/Runtime/Interfaces/IScriptLocator.cs
@@ -1,7 +1,0 @@
-﻿namespace ULox
-{
-    public interface IScriptLocator
-    {
-        Script Find(string name);
-    }
-}
