--- ulox/ulox.core/Package/Runtime/Interfaces/ITypeBodyCompilette.cs
@@ -1,7 +1,0 @@
-﻿namespace ULox
-{
-    public interface ITypeBodyCompilette : ICompilette
-    {
-        TypeCompiletteStage Stage { get; }
-    }
-}
