--- ulox/ulox.core/Package/Runtime/Interfaces/IULoxLibrary.cs
@@ -1,7 +1,0 @@
-﻿namespace ULox
-{
-    public interface IULoxLibrary
-    {
-        Table GetBindings();
-    }
-}
