--- ulox/ulox.core/Package/Runtime/Interfaces/IValueObjectBuilder.cs
@@ -1,10 +1,0 @@
-﻿namespace ULox
-{
-    public interface IValueObjectBuilder
-    {
-        IValueObjectBuilder CreateChild(string name);
-        IValueObjectBuilder CreateArray(string name);
-        Value Finish();
-        void SetField(string name, Value data);
-    }
-}
