+++ ulox/ulox.core/Package/Runtime/Library/Serialise/ValueHeirarchyWalker.cs
@@ -2,6 +2,17 @@
 
 namespace ULox
 {
+    public interface IValueHeirarchyWriter
+    {
+        void StartNamedElement(string name);
+        void StartElement();
+        void EndElement();
+        void StartArray(string name);
+        void EndArray();
+        void WriteNameAndValue(string name, Value v);
+        void WriteValue(Value v);
+    }
+    
     public class ValueHeirarchyWalker
     {
         private readonly IValueHeirarchyWriter _writer;
